requirements.txt
requests~=2.31.0
pytest
selene~=2.0.0rc9
allure-pytest